Move feeder into package and factor out datafeed
